+++ packages/react/src/components/Button/Default/index.tsx
@@ -1,3 +1,4 @@
+import { fontSizes, fontWeights, radii, space } from "@blue-health/tokens";
 import {
   Button as ButtonMUI,
   ButtonProps as MUIButtonProps,
@@ -13,11 +14,14 @@ export function ButtonDefault({ children, ...props }: IButtonDefaultProps) {
     <ButtonMUI
       {...props}
       sx={{
-        borderRadius: 1,
-        px: 2,
-        py: 1,
-        fontSize: 16,
-        fontWeight: 700,
+        borderRadius: radii.md,
+        px: space[4],
+        py: space[2],
+        borderStyle: "solid",
+        borderWidth: 2,
+        fontSize: fontSizes.md,
+        fontWeight: fontWeights.bold,
+        transition: "all 0.3s",
         textTransform: "none",
         width: "100%",
         ...props.sx,