+++ packages/react/src/components/TextField/index.ts
@@ -1,0 +1,13 @@
+import { TextFieldDefault, TextFieldDefaultProps } from "./Default";
+import { TextFieldExternalLabel } from "./ExternalLabel";
+import { TextFieldIcon } from "./Icon";
+import { TextFieldInternalLabel } from "./InternalLabel";
+
+export const TextField = {
+  Default: TextFieldDefault,
+  Icon: TextFieldIcon,
+  ExternalLabel: TextFieldExternalLabel,
+  InternalLabel: TextFieldInternalLabel,
+};
+
+export type TextFieldProps = TextFieldDefaultProps;