+++ packages/react/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @blue-health/react
 
+## 3.1.1
+
+### Patch Changes
+
+- Adjust components to use tokens
+
+## 3.1.0
+
+### Minor Changes
+
+- Create Text Field component
+
 ## 3.0.0
 
 ### Major Changes